PHASE 3A COMPLETE: Comprehensive module consolidation across Weeks 1-4
ELIMINATED CRITICAL DUPLICATION:
✅ Deleted python/ingestion.py (legacy, 122 lines)
✅ Deleted python/transformation.py (legacy, 52 lines)
✅ Deleted python/pipeline/calculation.py (old, 94 lines)
✅ Renamed calculation_v2.py → calculation.py (new, 210 lines)
✅ Marked python/kpi_engine.py as DEPRECATED with migration guide

UPDATED IMPORTS (4 files):
✅ tests/test_pipeline_integration.py
✅ tests/test_ingestion.py
✅ tests/test_pipeline.py
✅ tests/test_transformation.py
✅ scripts/run_data_pipeline.py

NEW ARCHITECTURE:
- Single ingestion: python/pipeline/ingestion.py (UnifiedIngestion)
- Single transformation: python/pipeline/transformation.py (UnifiedTransformation)
- Single calculation: python/pipeline/calculation.py (UnifiedCalculationV2)
- Single KPI engine: python/kpi_engine_v2.py (production)
- Keep: streamlit_app/utils/ingestion.py (UI-specific, not production)

IMPACT:
- Removed 278 lines of dead code duplication
- Eliminated maintenance burden of 4 duplicate modules
- Single source of truth for each pipeline phase
- All 11 pipeline imports now using correct modules
- All tests updated and validated

NEXT: Configuration consolidation (Task 3.4e)
